Updated utility functions to use the heap rather than static variables for better thread safety (e.g. fully_qualified_name).   Lots more work on the python port. Ignore-this: 6dcba66af710ef442465d23489368bd4

+++ include/misc.h
@@ -42,7 +42,7 @@ char *normalise_url(char *url);
 
 #if 0
 #define DEBUG(x, ...) printf("%s:%d %d: " x, \
-			     __FUNCTION__, __LINE__, pthread_self(), ## __VA_ARGS__);
+			     __FUNCTION__, __LINE__, (int)pthread_self(), ## __VA_ARGS__);
 #else
 #define DEBUG(x, ...)
 #endif
